download conversion data from firebase

+++ app/src/main/java/com/wanlok/calculator/model/Calculation.kt
@@ -3,6 +3,6 @@ package com.wanlok.calculator.model
 import java.util.Date
 
 data class Calculation(
-    var calculationId: Int,
+    var calculationId: Long,
     var dateTime: Date
 )
